+++ DataElaborator/ThreadedResultSimulator.h
@@ -43,22 +43,22 @@ protected:
     std::vector<ExperimentSimulator*> simulators;
     
     std::mt19937 rng;
-	std::uniform_real_distribution<double>* f1PosRangeDistribution;
-    std::uniform_real_distribution<double>* f2DistanceDistribution;
-    std::uniform_real_distribution<double>* f1apertureDistribution;
-    std::uniform_real_distribution<double>* f2apertureDistribution;
+	std::uniform_real_distribution<long double>* f1PosRangeDistribution;
+    std::uniform_real_distribution<long double>* f2DistanceDistribution;
+    std::uniform_real_distribution<long double>* f1apertureDistribution;
+    std::uniform_real_distribution<long double>* f2apertureDistribution;
     
     int simDataN = 4;
     int resultsN = 2;
     
     /* ------------------------  Simulation setup data ---------------------- */
-    double f1PosRange = 0.02;
-    double f2DistRange = 0.5;
-    double aperture = 7;
+    long double f1PosRange = 0.02;
+    long double f2DistRange = 0.5;
+    long double aperture = 7;
     int fenditureN=2;
     
     void SetupArrayResults();
-    double * results;
+    long double * results;
     int lastId;
 
 };
